strengthen the password hash
The fact that we were stripping characters between hashes at all
meant that we were actually weakening the crypto, and creating
the possibility of collisions.

By encoding the result from sha with base64, we make sure we're
maintaining the crypto strength, not creating collision issues, and
avoiding problems with invalid bytes as NULL (amongst others we were
not even filtering).

This solution is close to the recommendation for such cases on
python's bcrypt package itself:
https://github.com/pyca/bcrypt/blame/master/README.rst#L142